byte to int

+++ AnalyzeTrafficLight/AnalyzedObject.cs
@@ -20,7 +20,7 @@ namespace AnalyzeTrafficLight
 
     public class AnalyzedObject
     {
-        public byte id = 0;
+        public int id = 0;
 
         public ObjPoint leftTop = new ObjPoint();
